+++ packages/react-icons/build-verify.test.js
@@ -1728,7 +1728,7 @@ describe('Build Verification', () => {
         const jsContent = await readFile(jsFile, 'utf8');
         const trimmedJSContent = trimContentForSnapshot(jsContent);
         expect(trimmedJSContent).toMatchInlineSnapshot(`
-          "... (12 export void 0 declarations filtered (exports.Icon1 = exports.Icon3 = void 0))
+          "... (13 export void 0 declarations filtered (exports.Icon1 = exports.Icon3 = void 0))
           "use client";
           "use strict";
           Object.defineProperty(exports, "__esModule", { value: true });
@@ -1914,7 +1914,7 @@ describe('Build Verification', () => {
         const jsContent = await readFile(jsFile, 'utf8');
         const trimmedJSContent = trimContentForSnapshot(jsContent);
         expect(trimmedJSContent).toMatchInlineSnapshot(`
-          "... (44 export void 0 declarations filtered (exports.Icon1 = exports.Icon3 = void 0))
+          "... (45 export void 0 declarations filtered (exports.Icon1 = exports.Icon3 = void 0))
           "use client";
           "use strict";
           Object.defineProperty(exports, "__esModule", { value: true });
@@ -2084,15 +2084,15 @@ describe('Build Verification', () => {
       const { svgPathCjs, svgPathEsm } = getAssetPaths();
       const esmStats = await getAtomDirStats(svgPathEsm);
       const cjsStats = await getAtomDirStats(svgPathCjs, 'lib-cjs');
-      expect(esmStats.jsFiles.length).toMatchInlineSnapshot(`2896`);
-      expect(cjsStats.jsFiles.length).toMatchInlineSnapshot(`2896`);
+      expect(esmStats.jsFiles.length).toMatchInlineSnapshot(`2902`);
+      expect(cjsStats.jsFiles.length).toMatchInlineSnapshot(`2902`);
     });
     it(`should have same number of atoms/fonts icon files in lib and lib-cjs`, async () => {
       const { fontsPathCjs, fontsPathEsm } = getAssetPaths();
       const esmStats = await getAtomDirStats(fontsPathEsm);
       const cjsStats = await getAtomDirStats(fontsPathCjs, 'lib-cjs');
-      expect(esmStats.jsFiles.length).toMatchInlineSnapshot(`2889`);
-      expect(cjsStats.jsFiles.length).toMatchInlineSnapshot(`2889`);
+      expect(esmStats.jsFiles.length).toMatchInlineSnapshot(`2895`);
+      expect(cjsStats.jsFiles.length).toMatchInlineSnapshot(`2895`);
     });
     it.each(['lib', 'lib-cjs'])('should have atoms/svg directory with icon files in %s', async (libDir) => {
       const atomsSvgPath = path.join(__dirname, libDir, 'atoms', 'svg');