+++ packages/react-icons/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 2.0.339 (2026-08-26)
+
+### 🚀 Features
+
+- **react-icons-font-subsetting-webpack-plugin:** support rspack ([#1219](https://github.com/microsoft/fluentui-system-icons/pull/1219))
+
+## 2.0.338 (2026-08-21)
+
+### 🚀 Features
+
+- **react-icons:** add new and updated icons ([#1222](https://github.com/microsoft/fluentui-system-icons/pull/1222))
+
+### 🩹 Fixes
+
+- **react-icons:** render Light font icons ([#1217](https://github.com/microsoft/fluentui-system-icons/pull/1217))
+
 ## 2.0.337 (2026-08-13)
 
 This release contains icon updates